requirements.txt
Pillow==10.4.0
python-dotenv==1.0.1
streamlit==1.38.0
